feat: introduce GameConstants, CombatService, GenerationService, InteractionService, and MovementRulesService; refactor GameService and TurnAndAiService to utilize new services for improved game mechanics and interaction handling
