+++ .github/copilot-instructions.md
@@ -94,9 +94,25 @@ CIの想定シーケンス例: format → lint → type check。
 - 依存方向の逆流、UI へのビジネスロジック混在、直接 DOM 操作の埋め込み。
 - `any`、未使用コードの残置、色の直書き、任意 CSS の直付け、`<style>` タグ使用。
 
+## サービス層の構成
+
+詳細は [`docs/ARCHITECTURE.md`](../docs/ARCHITECTURE.md) を参照。
+
+| Service                  | 責務                                                        |
+| ------------------------ | ----------------------------------------------------------- |
+| **GameService**          | ファサード。初期化と各サブサービスへの委譲                  |
+| **InteractionService**   | 盤面操作: 駒選択、パネルクリック、ハイライト計算            |
+| **MovementRulesService** | 読み取り専用の移動ルール判定 (`canMoveTo`, `canCancelMove`) |
+| **GenerationService**    | ユニット生成: 最適パネル検索、駒生成、リソース消費          |
+| **CombatService**        | 戦闘解決: 駒 vs 駒、駒 vs 城壁                              |
+| **PieceService**         | 駒 CRUD、移動実行、パッシブ成長                             |
+| **PanelService**         | パネル初期化、隣接検索、状態クリア                          |
+| **TurnAndAiService**     | ターン進行、リソース加算、AI 制御                           |
+| **VictoryService**       | 勝利判定 (本拠地占領)                                       |
+| **BoardLayoutService**   | 六角形座標計算                                              |
+
 ## 将来の改善候補
 
-- 駒の移動処理に問題が起きました。直前に選択した駒の移動候補が後に選択した駒でも表示されており、実際に移動が可能になっています。
 - 背面攻撃ボーナスを実装する際に、駒の向きを管理する仕組みが必要です。
 - マスに値を付与するには駒が立ち止まっている必要があります。
 - 基本用語の定義をこのファイルに追加してください。
