+++ src/lib/domain/constants/GameConstants.ts
@@ -1,0 +1,13 @@
+// Board initialization
+export const HOME_BASE_INIT_RESOURCE = 5;
+export const HOME_BASE_INIT_CASTLE = 10;
+
+// Passive growth caps (per-turn gain ceiling, does NOT reduce existing values above cap)
+export const PASSIVE_RESOURCE_CAP = 5;
+export const PASSIVE_CASTLE_CAP = 5;
+
+// Unit generation
+export const RESOURCE_THRESHOLD_FOR_GENERATION = 5;
+
+// Movement
+export const DEFAULT_MAX_PIECES_PER_PANEL = 2;
